requirements.txt
python-telegram-bot>=21.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
jieba>=0.42.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
